+++ trunk/src/codigo/TP taller/TP taller/Vista/Dibujables/Particula.h
@@ -7,7 +7,7 @@ public:
 	~Particula(void);
 	SDL_Renderer* ren;
 	void actualizar();
-	void dibujar(SDL_Renderer* renderer, int corrimientoX, int corrimientoY, int escalaZoom, int anchoPx, int altoPx);
+	void dibujar(SDL_Renderer* renderer, int corrimientoX, int corrimientoY, float escalaZoom, int anchoPx, int altoPx);
 	int tiempoDeVida;
 private:
